Move workflow-compile logic to legacy module
